+++ packages/ui/Data/DataBind.ts
@@ -12,7 +12,7 @@ export interface DataBindProps extends BaseProps {
 
 /**
  * DataBind class.
- * @link https://ui.studiometa.dev/-/components/DataBind/
+ * @link https://ui.studiometa.dev/components/DataBind/
  */
 export class DataBind<T extends BaseProps = BaseProps> extends withGroup(Base, 'data:')<DataBindProps & T> {
   static config: BaseConfig = {