+++ packages/ui/Sentinel/Sentinel.ts
@@ -3,7 +3,7 @@ import type { BaseConfig } from '@studiometa/js-toolkit';
 
 /**
  * Sentinel class.
- * @link https://ui.studiometa.dev/-/components/Sentinel/
+ * @link https://ui.studiometa.dev/components/Sentinel/
  */
 export class Sentinel extends withIntersectionObserver(Base, { threshold: [0, 1] }) {
   /**